Support displayAlign and multiRowAlign.
In order to support displayAlign and multiRowAlign, creation of image
layers needs to be postponed until the height and width of all the
blocks is known. At that point, the offsets at which the images need to
be overlaid can be calculated according to these sizes and the values of
displayAlign and multiRowAlign.