Improve vertical text positioning.
When pango renders lines using some fonts (e.g., TiresiasScreenfont)
there is blankspace above the ascenders and below the descenders; other
fonts (e.g., DejaVu Sans) are rendered with space above the ascenders,
but none below the descenders. With the previous method of positioning
text this meant that text would be rendered differently depending on the
font. This patch provides and initial fix for this.