Change GstTtmlRender from base class to element.
The code was originally derived from textoverlay, in which there is a
base class (gstbasetextoverlay) from which child element classes
(e.g., gsttextoverlay) inherited. However, since in our case there is a
single child that doesn't add any functionality to the base class, we'll
make the base class into a renderer element in its own right.